my_project/module_64/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.116.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.2.10
fsspec==2025.9.0
h11==0.16.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
onnx==1.19.0
onnxruntime==1.22.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
protobuf==6.32.1
pydantic==2.11.8
pydantic_core==2.33.2
pyreadline3==3.5.4
python-multipart==0.0.20
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
torch==2.8.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
Werkzeug==3.1.3
